Consolidated RoboCommandParserFunctions into RoboCommandParser
The RoboCommandParserFunctions file has been deleted, with any methods being relocated per the following:
- General StringBuilder extensions now reside within ExtensionMethods.cs in the StringBuilderExtensions class.
- All Parser-Specific methods are now part of RoboCommandParser.cs
